Rename to heron-explorer.
