+++ heron/explorer/src/python/show.py
@@ -16,9 +16,9 @@ import logging
 import tornado.gen
 import tornado.ioloop
 from heron.ui.src.python.handlers.access import heron as API
-import heron.explore.src.python.args as args
-from heron.explore.src.python.physicalplan import get_topology_info
-import heron.explore.src.python.help as help
+import heron.explorer.src.python.args as args
+from heron.explorer.src.python.physicalplan import get_topology_info
+import heron.explorer.src.python.help as help
 from tabulate import tabulate
 import json
 import utils
