+++ scripts/packages/tools_template_bin.sh
@@ -129,13 +129,13 @@ fi
 if [ -L "${bin}/heron-ui" ]; then
   rm -f "${bin}/heron-ui"
 fi
-if [ -L "${bin}/heron-explore" ]; then
-  rm -f "${bin}/heron-explore"
+if [ -L "${bin}/heron-explorer" ]; then
+  rm -f "${bin}/heron-explorer"
 fi
 if [ -d "${base}" -a -x "${base}/bin/heron-tracker" ]; then
   rm -fr "${base}"
 fi
-if [ -d "${base}" -a -x "${base}/bin/heron-explore" ]; then
+if [ -d "${base}" -a -x "${base}/bin/heron-explorer" ]; then
   rm -fr "${base}"
 fi
 
@@ -154,7 +154,7 @@ echo -n .
 
 ln -s "${base}/bin/heron-tracker" "${bin}/heron-tracker"
 ln -s "${base}/bin/heron-ui"      "${bin}/heron-ui"
-ln -s "${base}/bin/heron-explore" "${bin}/heron-explore"
+ln -s "${base}/bin/heron-explorer" "${bin}/heron-explorer"
 echo -n .
 
 rm "${base}/heron-tools.tar.gz"
