+++ tools/rules/heron_tools.bzl
@@ -8,7 +8,7 @@ def heron_tools_bin_files():
     return [
         "//heron/tracker/src/python:heron-tracker",
         "//heron/ui/src/python:heron-ui",
-        "//heron/explore/src/python:heron-explore",
+        "//heron/explorer/src/python:heron-explorer",
     ]
 
 def heron_tools_conf_files():
